refactor: refactor the code structure of offline alignment
