+++ padiff/interfaces/__init__.py
@@ -16,5 +16,5 @@
 
 from .interfaces import create_model, assign_weight, auto_diff, check_dataloader
 from ..checker import check_report, check_params, check_grads, check_weights
-from ..dump_tools import set_dump_root_path, get_dump_root_path
+from ..tools import set_dump_root_path, get_dump_root_path
 from ..weight_init.special_init import add_special_init
