+++ padiff/interfaces/diff_utils.py
@@ -14,7 +14,7 @@
 
 from ..utils import for_each_tensor, log
 from ..guards import SyncStepGuard
-from ..dump_tools import dump_report, dump_weights, dump_grads
+from ..tools import dump_report, dump_weights, dump_grads
 from ..checker import check_report, check_weights, check_grads, global_compare_configs
 
 import paddle
